python-api/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
diffusers==0.35.2
dotenv==0.9.9
fastapi==0.104.1
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
packaging==25.0
Pillow==10.1.0
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
setuptools==70.2.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.27.0
sympy==1.14.0
tokenizers==0.22.1
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
transformers==4.57.1
triton==3.3.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.24.0
zipp==3.23.0
